+++ examples/kafkajs/admin/list-groups.js
@@ -1,5 +1,5 @@
 // require('kafkajs') is replaced with require('@confluentinc/kafka-javascript').KafkaJS.
-const { Kafka, ConsumerGroupStates } = require('@confluentinc/kafka-javascript').KafkaJS;
+const { Kafka, ConsumerGroupStates, ConsumerGroupTypes } = require('@confluentinc/kafka-javascript').KafkaJS;
 const { parseArgs } = require('node:util');
 
 async function adminStart() {
@@ -20,13 +20,20 @@ async function adminStart() {
         short: 's',
         multiple: true,
         default: [],
-      }
+      },
+      'types': {
+        type: 'string',
+        short: 't',
+        multiple: true,
+        default: [],
+      },
     },
   });
 
   let {
     'bootstrap-servers': bootstrapServers,
     states: matchConsumerGroupStates,
+    types: matchConsumerGroupTypes,
     timeout,
   } = args.values;
 
@@ -35,6 +42,9 @@ async function adminStart() {
   }
   matchConsumerGroupStates = matchConsumerGroupStates.map(
     state => ConsumerGroupStates[state]);
+  
+  matchConsumerGroupTypes = matchConsumerGroupTypes.map(
+    type => ConsumerGroupTypes[type]);
   
   const kafka = new Kafka({
     kafkaJS: {
@@ -48,13 +58,15 @@ async function adminStart() {
   try {
     const groupOverview = await admin.listGroups({
       timeout,
-      matchConsumerGroupStates
+      matchConsumerGroupStates,
+      matchConsumerGroupTypes
     });
     for (const group of groupOverview.groups) {
       console.log(`Group id: ${group.groupId}`);
       console.log(`\tType: ${group.protocolType}`);
       console.log(`\tIs simple: ${group.isSimpleConsumerGroup}`);
       console.log(`\tState: ${group.state}`);
+      console.log(`\tType: ${group.type}`);
     }
   } catch(err) {
     console.log('List topics failed', err);